Use existing Unity installation if found (#2)
Updates the cleanup check

+++ src/main/groovy/com/wooga/spock/extensions/uvm/interceptor/UnityInstallationFeatureInterceptor.groovy
@@ -110,6 +110,8 @@ class UnityInstallationFeatureInterceptor extends UnityInstallationManagingInter
 
     @Override
     void deleteInstallation(IMethodInvocation invocation) {
-        installation.location.deleteDir()
+        if(shouldClean) {
+            installation.location.deleteDir()
+        }
     }
 }
